requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
autobahn==22.7.1
Automat==20.2.0
channels==3.0.5
channels-redis==3.4.1
cloudinary==1.29.0
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==3.2.15
django-allauth==0.50.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
incremental==21.3.0
msgpack==1.0.4
oauthlib==3.2.1
Pillow==9.2.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.5.0
pyOpenSSL==22.1.0
python3-openid==3.2.0
pytz==2022.2.1
requests-oauthlib==1.3.1
service-identity==21.1.0
sqlparse==0.4.2
Twisted==22.8.0
txaio==22.2.1
types-cryptography==3.3.23
zope.interface==5.4.0
